+++ code/client/network.c
@@ -720,7 +720,7 @@ void NetConn_Send(Connection *conn)
 
     mbedtls_arc4_crypt(&conn->encrypt, size, buff, buff);
 
-    int result = send(conn->fd.handle, cast(const char *)out->data, out->size, 0);
+    int result = send(conn->fd.handle, cast(const char *)out->data, (int)out->size, 0);
     if (result == SOCKET_ERROR) {
         LogError("send failed: %d", os_errno);
         NetConn_HardShutdown(conn);
@@ -742,7 +742,7 @@ void NetConn_Recv(Connection *conn)
 
     uint8_t buffer[5840];
     size_t size = conn->in.capacity - conn->in.size;
-    int iresult = recv(conn->fd.handle, cast(char *)buffer, size, 0);
+    int iresult = recv(conn->fd.handle, cast(char *)buffer, (int)size, 0);
 
     int err = os_errno;
     if (iresult == SOCKET_ERROR) {
@@ -987,7 +987,7 @@ static int pack(const uint8_t *data, size_t data_size,
                 memrcpy(wpos, rpos, bytes_to_write);
             #endif
 
-            readed  += static_elem_count * elem_size;
+            readed  += (int)(static_elem_count * elem_size);
             written += bytes_to_write;
         } else { // field.type == TYPE_NESTED_STRUCT
             assert(i + 1 < fields_count);
@@ -997,10 +997,10 @@ static int pack(const uint8_t *data, size_t data_size,
 
             for (size_t j = 0; j < packed_elem_count; j++) {
                 const uint8_t *d = data + readed;
-                int d_size = data_size - readed;
+                size_t d_size = data_size - readed;
 
                 uint8_t *b = buffer + written;
-                int b_size = buff_size - written;
+                size_t b_size = buff_size - written;
 
                 int tmp = pack(d, d_size, b, b_size, struct_fields, struct_fields_count);
                 if (tmp < 0) return -1;
@@ -1009,11 +1009,11 @@ static int pack(const uint8_t *data, size_t data_size,
             }
 
             // readed += field.size * remaining_count;
-            return written;
+            return (int)written;
         }
     }
 
-    return written;
+    return (int)written;
 }
 
 static int unpack(const uint8_t *data, size_t data_size, uint8_t *buffer,
@@ -1063,14 +1063,14 @@ static int unpack(const uint8_t *data, size_t data_size, uint8_t *buffer,
             static_elem_count = field.param;
             if (field.type == TYPE_STRING_16) {
                 // We want to ensure last character is null.
-                uint32_t end = MIN(field.param - 1, packed_elem_count);
+                uint32_t end = MIN(field.param - 1, (uint32_t)packed_elem_count);
                 s[end] = 0;
             } else {
                 written += sizeof(struct array);
-                a->size = packed_elem_count;
+                a->size = (uint32_t)packed_elem_count;
             }
             
-            readed += prefix_size;
+            readed += (int)prefix_size;
             // written += sizeof(array);
 
             rpos = data   + readed;
@@ -1093,19 +1093,19 @@ static int unpack(const uint8_t *data, size_t data_size, uint8_t *buffer,
                 memrcpy(wpos, rpos, bytes_to_read);
             #endif
 
-            readed  += bytes_to_read;
-            written += static_elem_count * elem_size;
+            readed  += (int)bytes_to_read;
+            written += (int)(static_elem_count * elem_size);
         } else { // field.type == TYPE_NESTED_STRUCT
             assert(i + 1 < fields_count);
             MsgField *next_fields = fields + i + 1;
-            int next_fields_count = fields_count - i - 1;
+            int next_fields_count = (int)(fields_count - i - 1);
 
             for (size_t j = 0; j < packed_elem_count; j++) {
                 const uint8_t *d = data + readed;
-                int d_size = data_size - readed;
+                int d_size = (int)data_size - readed;
 
                 uint8_t *b = buffer + written;
-                int b_size = buff_size - written;
+                int b_size = (int)buff_size - written;
 
                 int tmp = unpack(d, d_size, b, b_size, next_fields, next_fields_count);
                 if (tmp < 0) return -1;
