+++ code/portal/login.c
@@ -257,7 +257,7 @@ static int read_user_code(char *buffer, size_t size, size_t *ret)
 {
     printf("Enter code: ");
 
-    if (fgets(buffer, size, stdin) == NULL)
+    if (fgets(buffer, (int)size, stdin) == NULL)
         return 1;
 
     // The complete line didn't fit in the buffer.
